requirements.txt
sentence-transformers==2.2.2
torch>=2.6.0,<3.0.0
numpy==1.23.5
psutil==5.9.5
GPUtil==1.4.0
llama-cpp-python>=0.1.97
PyPDF2==3.0.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.2.0
ebooklib==0.18
beautifulsoup4==4.12.3
rich==13.7.1
loguru==0.7.2
typing-extensions==4.5.0
pathlib2==2.3.7.post1
